Move implementations out of .h-files
Separate the implementations of functions from the header-files
in ./src/plugins/hidsim/gui/*.h and move those into corresponding
.c-files to build a cleaner interface
