+++ src/plugins/hidsim/gui/vmi_win_gui_atoms.cpp
@@ -104,31 +104,204 @@
  * It is distributed as part of DRAKVUF under the same license             *
  ***************************************************************************/
 
-#ifndef VMI_WIN_GUI_UTILS_H
-#define VMI_WIN_GUI_UTILS_H
+#include <stdio.h>
 
-#include <wchar.h>
-#include <libvmi/libvmi.h>
+/* For PRINT_DEBUG */
+#include "../../private.h"
+/* For eprint_current_time  */
+#include <libdrakvuf/libdrakvuf.h>
+
 #include "vmi_win_gui_offsets.h"
-
-#define ARRAY_SIZE(x) (sizeof(x) / sizeof((x)[0]))
+#include "vmi_win_gui_utils.h"
 
 extern struct Offsets symbol_offsets;
 
-/*
- * Checks, whether the _OBJECT_HEADER, which precedes every executive object,
- * is preceded by an optional header of type _OBJECT_HEADER_NAME_INFO.
- * If this is the case, the name of the executive object is read and returned.
- *
- */
-char* retrieve_objhdr_name(vmi_instance_t vmi, addr_t addr);
-/*
- * Reads a Windows wchar-string into a wchar_t*, since vmi_read_unicode_str_va
- * fails to parse _RTL_ATOM_ENTRY's name-string or _LARGE_UNICODE_STRINGs.
- * Expansion is performed since Windows' wchar is 2 bytes versus 4 bytes on
- * 64bit-Linux
- */
-wchar_t* read_wchar_str_pid(vmi_instance_t vmi, addr_t start, size_t len, vmi_pid_t pid);
+struct atom_entry
+{
+    uint16_t atom;
+    uint16_t ref_count;
+    addr_t hashlink;
+    uint8_t name_len;
+    wchar_t* name;
+};
 
-wchar_t* read_wchar_str(vmi_instance_t vmi, addr_t start, size_t len);
-#endif // VMI_WIN_GUI_UTILS_H
+void free_atom_entry(struct atom_entry* a)
+{
+    if (a)
+        free(a->name);
+
+    free(a);
+}
+
+/* Prints a single atom entry */
+void print_atom(gpointer key, gpointer value, gpointer user_data )
+{
+    PRINT_DEBUG("Atom: %" PRIx16 " %ls\n", GPOINTER_TO_UINT(key),
+        ((struct atom_entry*)value)->name);
+}
+
+/*
+ * Parses the in memory kernel data structure representing an atom entry and
+ * stores the result in a struct of type atom_entry
+ */
+struct atom_entry* parse_atom_entry(vmi_instance_t vmi, addr_t atom_addr)
+{
+    struct atom_entry* entry = (struct atom_entry*) malloc(sizeof(struct atom_entry));
+    memset(entry, 0, sizeof(struct atom_entry));
+
+    if (VMI_FAILURE == vmi_read_addr_va(vmi, atom_addr +
+            symbol_offsets.atom_entry_hashlink_offset, 0, &entry->hashlink))
+    {
+        printf("Error reading HashLink at %" PRIx64 "\n", atom_addr +
+            symbol_offsets.atom_entry_hashlink_offset);
+        return NULL;
+    }
+
+    if (VMI_FAILURE == vmi_read_16_va(vmi, atom_addr +
+            symbol_offsets.atom_entry_atom_offset, 0, (uint16_t*)&entry->atom))
+    {
+        printf("Error reading Atom at %" PRIx64 "\n", atom_addr +
+            symbol_offsets.atom_entry_atom_offset);
+        return NULL;
+    }
+
+    if (VMI_FAILURE == vmi_read_16_va(vmi, atom_addr +
+            symbol_offsets.atom_entry_ref_count_offset, 0, (uint16_t*)&entry->ref_count))
+    {
+        printf("Error reading ReferenceCount at %" PRIx64 "\n", atom_addr +
+            symbol_offsets.atom_entry_ref_count_offset);
+        return NULL;
+    }
+
+    if (VMI_FAILURE == vmi_read_8_va(vmi, atom_addr +
+            symbol_offsets.atom_entry_name_len_offset, 0, (uint8_t*)&entry->name_len))
+    {
+        printf("Error reading NameLength at %" PRIx64 "\n", atom_addr +
+            symbol_offsets.atom_entry_name_len_offset);
+        return NULL;
+    }
+
+    entry->name = read_wchar_str(vmi, atom_addr +
+            symbol_offsets.atom_entry_name_offset, (size_t)entry->name_len);
+
+    if (!entry->name)
+    {
+        printf("Error reading wchar-string Name at %" PRIx64 "\n", atom_addr +
+            symbol_offsets.atom_entry_name_offset);
+    }
+
+    return entry;
+}
+
+/* Creates a single atom-entry struct */
+struct atom_entry* create_atom_entry(uint16_t atom, const wchar_t* name,
+    uint8_t len, addr_t hashlink, uint16_t refcount)
+{
+    struct atom_entry* entry = (struct atom_entry*) malloc(sizeof(struct atom_entry));
+    memset(entry, 0, sizeof(struct atom_entry));
+    entry->atom = atom;
+    entry->name = wcsdup(name);
+    entry->name_len = len;
+    entry->hashlink = hashlink;
+    entry->ref_count = refcount;
+
+    return entry;
+}
+
+/*
+ * Adds the default _RTL_ATOM_ENTRY-structs to the atom-table
+ *
+ * Default _RTL_ATOM_ENTRY-structs
+ * See https://github.com/volatilityfoundation/volatility/blob/\
+ * a438e768194a9e05eb4d9ee9338b881c0fa25937/volatility/plugins/gui/\
+ * constants.py#L34
+ */
+void add_default_atoms(GHashTable* atom_table)
+{
+    struct atom_entry* ae = NULL;
+
+    ae = create_atom_entry(0x8000, L"PopupMenu", 9, 0, 0);
+    g_hash_table_insert(atom_table, GUINT_TO_POINTER(ae->atom), (gpointer)ae);
+
+    ae = create_atom_entry(0x8001, L"Desktop", 7, 0, 0);
+    g_hash_table_insert(atom_table, GUINT_TO_POINTER(ae->atom), (gpointer)ae);
+
+    ae = create_atom_entry(0x8002, L"Dialog", 6, 0, 0);
+    g_hash_table_insert(atom_table, GUINT_TO_POINTER(ae->atom), (gpointer)ae);
+
+    ae = create_atom_entry(0x8003, L"WinSwitch", 9, 0, 0);
+    g_hash_table_insert(atom_table, GUINT_TO_POINTER(ae->atom), (gpointer)ae);
+
+    ae = create_atom_entry(0x8004, L"IconTitle", 9, 0, 0);
+    g_hash_table_insert(atom_table, GUINT_TO_POINTER(ae->atom), (gpointer)ae);
+
+    ae = create_atom_entry(0x8006, L"ToolTip", 9, 0, 0);
+    g_hash_table_insert(atom_table, GUINT_TO_POINTER(ae->atom), (gpointer)ae);
+
+}
+
+/*
+ * Builds up the atom table, which serves as a shared resource hosting class
+ * IDs and their name (regularily used by more than one process)
+ *
+ * Additional background information can be found at:
+ * https://bsodtutorials.wordpress.com/2015/11/11/understanding-atom-tables/
+ */
+GHashTable* build_atom_table(vmi_instance_t vmi, addr_t table_addr)
+{
+    uint32_t num_buckets = 0;
+
+    if (VMI_FAILURE == vmi_read_32_va(vmi, table_addr +
+            symbol_offsets.atom_table_num_buckets_off, 0, (uint32_t*)&num_buckets))
+    {
+        printf("Failed to read num buckets-value of _RTL_ATOM_TABLE at %" PRIx64
+            "\n", table_addr + symbol_offsets.atom_table_num_buckets_off);
+        return NULL;
+    }
+
+    GHashTable* ht = g_hash_table_new_full(g_direct_hash, g_direct_equal, NULL,
+            (GDestroyNotify) free_atom_entry);
+    add_default_atoms(ht);
+
+    size_t i = 0;
+    addr_t cur = 0;
+    struct atom_entry* a = NULL;
+
+    /* Iterate the array of pointers to _RTL_ATOM_TABLE_ENTRY-structs at buckets */
+    while (i < num_buckets)
+    {
+        if (VMI_FAILURE == vmi_read_addr_va(vmi, table_addr +
+                symbol_offsets.atom_table_buckets_off + i * 4, 0, &cur))
+        {
+            printf("Failed to read pointer to buckets entry of _RTL_ATOM_TABLE at %"
+                PRIx64 "\n", table_addr + symbol_offsets.atom_table_buckets_off + i * 4);
+            g_hash_table_destroy(ht);
+            return NULL;
+        }
+        i++;
+
+        if (!cur)
+            continue;
+
+        a = parse_atom_entry(vmi, cur);
+
+        if (a)
+        {
+            g_hash_table_insert(ht, GUINT_TO_POINTER(a->atom), (gpointer)a);
+        }
+
+        /* Traverses the linked list of each top level _RTL_ATOM_TABLE_ENTRY */
+        while (a && a->hashlink)
+        {
+            cur = a->hashlink;
+            a = parse_atom_entry(vmi, cur);
+
+            if (a)
+            {
+                g_hash_table_insert(ht, GUINT_TO_POINTER(a->atom), (gpointer)a);
+            }
+        }
+    }
+
+    return ht;
+}
